fix(accounts): revoke tenant access on member removal

+++ app/models/user.rb
@@ -109,7 +109,13 @@ class User < ApplicationRecord
     return false unless membership
     return false unless normalize_role(role, resource) == membership.role
 
-    membership.destroy
+    case resource
+    when Account
+      revoke_account_access!(resource)
+    else
+      membership.destroy
+    end
+
     true
   end
 
@@ -165,6 +171,28 @@ class User < ApplicationRecord
 
   def operator?
     OperatorConsole::Access.allowed?(self)
+  end
+
+  def revoke_account_access!(account)
+    with_lock do
+      membership = account_memberships.find_by(account: account)
+      return false unless membership
+
+      if account_id == account.id
+        successor_membership = account_memberships.where.not(id: membership.id).order(:id).first
+
+        if successor_membership
+          update!(account: successor_membership.account)
+          membership.destroy!
+        else
+          destroy!
+        end
+      else
+        membership.destroy!
+      end
+    end
+
+    true
   end
 
   private
